VArify/requirements.txt
astroid==2.4.2
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
dnspython==2.0.0
Flask==1.1.2
Flask-PyMongo==2.3.0
Flask-WTF==0.14.3
geographiclib==1.50
geopy==2.0.0
gunicorn==20.0.4
idna==2.10
iniconfig==1.1.1
isort==5.4.2
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.3.1
mccabe==0.6.1
numpy==1.19.1
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
py==1.9.0
pylint==2.6.0
pymongo==3.11.0
pyparsing==2.4.7
pytest==6.1.1
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
six==1.15.0
toml==0.10.1
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3